+++ CHANGELOG.md
@@ -45,7 +45,7 @@ ADDED:
 - Ability to define options for `JsonApiClient`
 - `Document::error($number)` utility method to retrieve only one error
 - `Response::isSuccessful()` to determine if the response was successful
-- `Response::isSuccessful()` to determine if the response was successful and contains a document
+- `Response::isSuccessfulDocument()` to determine if the response was successful and contains a document
 - Support for PHPUnit 5.0
 
 CHANGED: