+++ CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGED:
 - Renamed `JsonApiAsyncClient::requestConcurrent()` to `JsonApiAsyncClient::sendConcurrentAsyncRequests()`
 - Renamed `Document::hasPrimaryResources()` to `Docment::hasAnyPrimaryResources()`
 - Renamed `Document::hasIncludedResources()` to `Docment::hasAnyIncludedResources()`
+- Renamed almost all methods in `JsonApiRequestBuilder`
 - Improved documentation
 
 REMOVED: